+++ lib/shelly/cli/main/add.rb
@@ -17,6 +17,8 @@ module Shelly
         :desc => "Skip Shelly Cloud requirements check"
       method_option "zone", :type => :string, :hide => true,
         :desc => "Create cloud in given zone"
+      method_option "region", :type => :string,
+        :desc => "Create cloud in given region"
       desc "add", "Add a new cloud"
       def add
         check_options(options)
@@ -27,8 +29,13 @@ module Shelly
         app.code_name = options["code-name"] || ask_for_code_name
         app.databases = options["databases"] || ask_for_databases
         app.size = options["size"] || "small"
-        app.organization_name = options["organization"] || ask_for_organization(options)
-        app.zone = options["zone"]
+        app.organization_name = options["organization"] ||
+          ask_for_organization(options)
+        if options["zone"].present?
+          app.zone = options["zone"]
+        else
+          app.region = options["region"] || ask_for_region
+        end
 
         app.create
         say "Cloud '#{app}' created in '#{app.organization_name}' organization", :green
@@ -60,7 +67,7 @@ module Shelly
         e.each_error { |error| say_error error, :with_exit => false }
         say_new_line
         say_error "Fix erros in the below command and type it again to create your cloud" , :with_exit => false
-        say_error "shelly add --code-name=#{app.code_name.downcase.dasherize} --databases=#{app.databases.join(',')} --organization=#{app.organization_name} --size=#{app.size}"
+        say_error "shelly add --code-name=#{app.code_name.downcase.dasherize} --databases=#{app.databases.join(',')} --organization=#{app.organization_name} --size=#{app.size} --region=#{app.region}"
       rescue Client::ForbiddenException
         say_error "You have to be the owner of '#{app.organization_name}' organization to add clouds"
       rescue Client::NotFoundException => e
@@ -72,33 +79,33 @@ module Shelly
       no_tasks do
         def ask_for_organization(options)
           organizations = Shelly::User.new.organizations
+
           if organizations.blank?
             ask_for_new_organization(options)
           else
-            count = organizations.count
-            option_selected = 0
+            say "Select organization for this cloud:"
+            say_new_line
+
             loop do
-              say "Select organization for this cloud:"
-              say_new_line
               say "existing organizations:"
 
               organizations.each_with_index do |organization, i|
                 print_wrapped "#{i + 1}) #{organization.name}", :ident => 2
               end
+
+              say green "Or leave empty to create a new organization"
               say_new_line
 
-              print_wrapped "#{count + 1}) provide name for new organization", :ident => 2
-
-              option_selected = ask("Option:")
-              break if ('1'..(count + 1).to_s).include?(option_selected)
-            end
-
-            if option_selected.to_i == count + 1
-              return ask_for_new_organization(options)
-            end
-
-            if (1..count).include?(option_selected.to_i)
-              return organizations[option_selected.to_i - 1].name
+              selected = ask("Organization:")
+              if organizations.select { |o| o.name == selected }.present?
+                return selected
+              elsif selected.empty?
+                say_new_line
+                return ask_for_new_organization(options)
+              else
+                say_new_line
+                say_warning "#{selected} organization does not exist"
+              end
             end
           end
         end
@@ -113,6 +120,28 @@ module Shelly
           end
         end
 
+        def ask_for_region
+          regions = Shelly::App::REGIONS
+          say "Select region for this cloud:"
+          say_new_line
+
+          loop do
+            say "available regions:"
+
+            regions.each_with_index do |region, i|
+              print_wrapped "#{i + 1}) #{region}", :ident => 2
+            end
+            say_new_line
+
+            selected = ask("Region:").upcase
+            if regions.include?(selected)
+              return selected
+            else
+              say_new_line
+              say_warning "#{selected} region is not available"
+            end
+          end
+        end
       end
     end
   end